did pattern matching

+++ internal/session/session.go
@@ -4,7 +4,7 @@ import "time"
 
 type Session struct {
 	SessionId  string    `json:"session_id"`
-	UserID     int       `json:"user_id"`
+	UserID     int64     `json:"user_id"`
 	CreatedAt  time.Time `json:"created_at"`
 	ValidUntil time.Time `json:"valid_until"`
 }
